backend/requirements.txt
fastapi
uvicorn
pydantic
openai
httpx
beautifulsoup4
python-multipart
python-dotenv
duckduckgo-search
